requirements.txt
yfinance>=0.2.28
snowflake-connector-python>=3.0.0
pandas>=2.0.0
pytz>=2023.3
python-dotenv>=1.0.0
requests>=2.31.0
